+++ zia/resource_zia_firewall_filtering_rules.go
@@ -279,6 +279,7 @@ func resourceFirewallFilteringRulesCreate(ctx context.Context, d *schema.Resourc
 	_ = d.Set("rule_id", resp.ID)
 
 	markOrderRuleAsDone(resp.ID, resourceType)
+	waitForReorder(resourceType)
 
 	// Check if ZIA_ACTIVATION is set to a truthy value before triggering activation
 	if shouldActivate() {
@@ -510,6 +511,7 @@ func resourceFirewallFilteringRulesUpdate(ctx context.Context, d *schema.Resourc
 		return diags
 	}
 	markOrderRuleAsDone(req.ID, "firewall_filtering_rules")
+	waitForReorder("firewall_filtering_rules")
 
 	// Sleep for 2 seconds before potentially triggering the activation
 	time.Sleep(2 * time.Second)